requirements.txt
certifi==2019.9.11
chardet==3.0.4
entrypoints==0.3
flake8==3.7.8
future==0.17.1
idna==2.8
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-decouple==3.1
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.6
